fix(ZNTA-1740): ensure UTF-8 when parsing TMX segments
perf: reuse XMLInputFactory and ensure XMLEventReader is closed
refactor: move TMX code to TMXUtil
